requirements.txt
confluent-kafka==2.8.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.6
faker==35.2.0
pytest==8.3.5